refactor: Reorg and rename pool test files.
