+++ test/fxPool.test.ts
@@ -1,7 +1,7 @@
 // pnpm test -- fxPool.test.ts
 import { describe, expect, test } from 'vitest';
 import { ChainId, RawFxPool, SwapKind, Token } from '../src';
-import testPools from './lib/testData/fxPool_43667355.json';
+import testPools from './lib/testData/testPools/fx_43667355.json';
 import { CurveMathRevert, FxPool, FxPoolToken } from '../src/entities/pools/fx';
 import { parseFixedCurveParam } from '../src/entities/pools/fx/helpers';
 import { parseUnits } from 'viem';
