+++ test/gyroEMath.test.ts
@@ -1,9 +1,10 @@
 // pnpm test -- gyroEMath.test.ts
 import { describe, expect, test } from 'vitest';
-import testPools from './lib/testData/gyroETestPool.json';
+import testPools from './lib/testData/testPools/gyroE_44215395.json';
 import { RawGyroEPool } from '../src/data/types';
 import { ChainId } from '../src/utils';
-import { GyroEPool, Vector2 } from '../src/entities/pools/gyroE/gyroEPool';
+import { GyroEPool } from '../src/entities/pools/gyroE/gyroEPool';
+import { Vector2 } from '../src/entities/pools/gyroE/types';
 import { Token, TokenAmount } from '../src/entities';
 import { calculateInvariantWithError } from '../src/entities/pools/gyroE/gyroEMath';
 import {
