+++ test/gyroEPool.test.ts
@@ -1,6 +1,6 @@
 // pnpm test -- test/gyroEPool.test.ts
 import { describe, expect, test } from 'vitest';
-import testPools from './lib/testData/gyroETestPool.json';
+import testPools from './lib/testData/testPools/gyroE_44215395.json';
 import { ChainId, SwapKind, Token, TokenAmount } from '../src';
 import { RawGyroEPool } from '../src/data/types';
 import { GyroEPool } from '../src/entities/pools/gyroE';
