+++ test/gyroEV2Pool.integration.test.ts
@@ -3,7 +3,7 @@ import { beforeAll, beforeEach, describe, expect, test } from 'vitest';
 import dotenv from 'dotenv';
 dotenv.config();
 
-import testPools from './lib/testData/gyroETestPool.json';
+import testPools from './lib/testData/testPools/gyroE_44215395.json';
 import { BATCHSIZE, ChainId, VAULT } from '../src/utils';
 import {
     BasePool,
